requirements.txt
asgiref==3.8.1
certifi==2025.1.31
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.1
Django==5.2
django-debug-toolbar==5.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
msgpack==1.1.0
mysqlclient==2.2.7
pillow==11.1.0
PyJWT==2.9.0
redis==5.2.1
requests==2.32.3
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.3.0
